requirements.txt
# Required packages for DSCI-560 Lab 4
# Stock Data Collection and Algorithmic Trading

# Data Collection
polygon-api-client==1.12.5
six>=1.16.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
# SQLite3 comes with Python, no need to install

# Data Visualization (for analysis)
matplotlib==3.8.2
seaborn==0.13.0

# Technical Analysis (for algorithms in next steps)
ta==0.11.0
pandas-ta==0.3.14b0

# Optional: For advanced algorithms
scikit-learn==1.3.2
statsmodels==0.14.1
